# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Portfolio API",
        "description": "This is the Swagger UI documentation for my Laravel API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/blog": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get all blog",
                "operationId": "ea1c9e4eb5fdd360cfe59cdefce55814",
                "responses": {
                    "200": {
                        "description": "List of blog"
                    }
                }
            },
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Create a new blog",
                "operationId": "1605fee407ba78ac717c1c883b630c17",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "b_title",
                                    "b_subtitle",
                                    "b_detail",
                                    "b_img",
                                    "b_date",
                                    "display"
                                ],
                                "properties": {
                                    "b_img": {
                                        "type": "integer"
                                    },
                                    "b_title": {
                                        "type": "string"
                                    },
                                    "b_subtitle": {
                                        "type": "string"
                                    },
                                    "b_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "b_detail": {
                                        "type": "string"
                                    },
                                    "display": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Blog created successfully"
                    }
                }
            }
        },
        "/api/blog/{id}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get an blog by ID",
                "operationId": "6ac9d2eb1f1ccbc64de4ae3a2bb02941",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blog retrieved successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Update a blog by ID",
                "operationId": "ebfc9751cbe46ca77c6c6f33efabc7ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "b_img": {
                                        "type": "integer"
                                    },
                                    "b_title": {
                                        "type": "string"
                                    },
                                    "b_subtitle": {
                                        "type": "string"
                                    },
                                    "b_detail": {
                                        "type": "string"
                                    },
                                    "b_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "display": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Blog updated successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                }
            }
        },
        "/api/blog/visible/{id}": {
            "put": {
                "tags": [
                    "Blog"
                ],
                "summary": "Delete an blog by ID",
                "operationId": "b73fcda3b8f8896c9d41de228bd7b626",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete blog successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                }
            }
        },
        "/api/blog/reorder": {
            "put": {
                "tags": [
                    "Blog"
                ],
                "summary": "Reorder blogs",
                "operationId": "06fd10c1d8fb3fa2a0e37939435b5cae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "b_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "b_order": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Blog order updated successfully"
                    }
                }
            }
        },
...
    },
...
}